clened debug in view
